+++ projects/summer-fi-pro/helpers/get-automation-cdp-ids.js
